Updated configure perl scripts to use configperl ref#29
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/core/trunk@41 77d55b08-817d-0410-979b-1fec1beb8c2e

+++ AbstractNamedObjectContainer.cpp
@@ -125,7 +125,7 @@ AbstractNamedObject *AbstractNamedObjectContainer::FindNamedObject(const char *n
     while (*name != 0) {
         next = ::strstr(name, "::");
         if (next != NULL) {
-            n = vislib::StringA(name, next - name);
+            n = vislib::StringA(name, static_cast<int>(next - name));
             name = next + 2;
         } else {
             n = name;
